requirements.txt
pandas
matplotlib
seaborn
streamlit
plaid-python
sqlalchemy
fastapi
uvicorn
python-dotenv
openai